+++ xrEngine/xrImage_Resampler.cpp
@@ -324,8 +324,8 @@ void imf_Process(u32 *dstI, u32 dstW, u32 dstH, u32 *srcI, u32 srcW, u32 srcH, E
 				contrib[i].p = (CONTRIB *)xr_malloc((int)(width * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(width * 2 + 1) * sizeof(CONTRIB));
 				center = float(i) / xscale;
-				left = ceil(center - width);
-				right = floor(center + width);
+				left = ceilf(center - width);
+				right = floorf(center + width);
 				for (j = int(left); j <= int(right); ++j)
 				{
 					weight = center - float(j);
@@ -363,8 +363,8 @@ void imf_Process(u32 *dstI, u32 dstW, u32 dstH, u32 *srcI, u32 srcW, u32 srcH, E
 				contrib[i].p = (CONTRIB *)xr_malloc((int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				center = float(i) / xscale;
-				left = ceil(center - fwidth);
-				right = floor(center + fwidth);
+				left = ceilf(center - fwidth);
+				right = floorf(center + fwidth);
 				for (j = int(left); j <= int(right); ++j)
 				{
 					weight = center - (float)j;
@@ -465,8 +465,8 @@ void imf_Process(u32 *dstI, u32 dstW, u32 dstH, u32 *srcI, u32 srcW, u32 srcH, E
 				contrib[i].p = (CONTRIB *)xr_malloc((int)(width * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(width * 2 + 1) * sizeof(CONTRIB));
 				center = (float)i / yscale;
-				left = ceil(center - width);
-				right = floor(center + width);
+				left = ceilf(center - width);
+				right = floorf(center + width);
 				for (j = int(left); j <= int(right); ++j)
 				{
 					weight = center - (float)j;
@@ -504,8 +504,8 @@ void imf_Process(u32 *dstI, u32 dstW, u32 dstH, u32 *srcI, u32 srcW, u32 srcH, E
 				contrib[i].p = (CONTRIB *)xr_malloc((int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				ZeroMemory(contrib[i].p, (int)(fwidth * 2 + 1) * sizeof(CONTRIB));
 				center = (float)i / yscale;
-				left = ceil(center - fwidth);
-				right = floor(center + fwidth);
+				left = ceilf(center - fwidth);
+				right = floorf(center + fwidth);
 				for (j = int(left); j <= int(right); ++j)
 				{
 					weight = center - (float)j;
